requirements.txt
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.1
django-cors-headers==2.4.0
django-recaptcha==1.4.0
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
idna==2.7
libgravatar==0.2.3
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
ua-parser==0.8.0
urllib3==1.23
user-agents==1.1.0
